Update  finance_calculators.py
Line Length: Lines are kept under 80 characters for better readability.
